Move drawer to parent component
