+++ client/src/views/Mapping/components/FhirMappingPanel/components/ResourceSelector/Drawer/index.test.tsx
