+++ client/src/views/Mapping/components/FhirMappingPanel/components/ResourceSelector/Drawer/style.less
